some specific things

+++ backend/redmine/serializers.py
@@ -1,4 +1,5 @@
 from rest_framework import serializers
 
+
 class RedmineExtraFieldPayloadSerializer(serializers.Serializer):
-    hours = serializers.DecimalField(max_digits=10, decimal_places=7)
+    hours = serializers.FloatField()
